refactor(client): move SettingsPanel.tsx into SettingsPanel/ directory
Resolve naming conflict where SettingsPanel.tsx (1,547 LOC) sat as a
sibling file next to the SettingsPanel/ directory containing its
sub-components. Move the file into the directory and update index.ts
to export from the local file instead of the parent. Fix relative
import paths that now need an additional level of traversal.
