+++ packages/client/src/game/panels/SettingsPanel/index.ts
@@ -6,8 +6,6 @@
  * @packageDocumentation
  */
 
-// Re-export the main panel from the original location
-// This allows gradual migration to the new structure
-export { SettingsPanel } from "../SettingsPanel";
+export { SettingsPanel } from "./SettingsPanel";
 export * from "./components";
 export * from "./constants";
